Issue #4538 - Combine MessageWriterTests, add test for UTF-8 validation
